@@ -14,7 +14,7 @@
     }
   ],
   "require": {
-    "php": "^7.4 || ^8.0",
+    "php": "^8.1",
     "typo3/cms-core": "^12.4",
     "typo3/cms-extbase": "^12.4",
     "typo3/cms-dashboard": "^12.4",
@@ -32,7 +32,8 @@
     "typo3/testing-framework": "^7.0.2",
     "helhum/typo3-console": "^8.0",
     "typo3/minimal": "^12.4",
-    "typo3/cms-lowlevel": "^12.4"
+    "typo3/cms-lowlevel": "^12.4",
+    "rector/rector": "^0.19.1"
   },
   "replace": {
     "typo3-ter/t3-messenger-dashboard": "self.version"
