@@ -13,11 +13,8 @@ namespace Ssch\T3MessengerDashboard\Tests\Functional\Fixtures\Extensions\t3_mess
 
 final class MyFailingCommand
 {
-    private string $note;
-
-    public function __construct(string $note)
+    public function __construct(private readonly string $note)
     {
-        $this->note = $note;
     }
 
     public function getNote(): string
