@@ -11,14 +11,15 @@ declare(strict_types=1);
 
 namespace Ssch\T3MessengerDashboard\Tests\Functional\Fixtures\Extensions\t3_messenger_dashboard_test\Classes\Handlers;
 
+use Ssch\T3MessengerDashboard\Tests\Functional\Fixtures\Extensions\t3_messenger_dashboard_test\Classes\Command\MyFailingCommand;
 use Ssch\T3MessengerDashboard\Tests\Functional\Fixtures\Extensions\t3_messenger_dashboard_test\Classes\Command\MyOtherFailingCommand;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
 
 #[AsMessageHandler]
-final class MyMessengerHandler
+final class MySecondMessengerHandler
 {
 
-    public function __invoke(MyOtherFailingCommand $command): never
+    public function __invoke(MyFailingCommand $command): never
     {
         throw new \InvalidArgumentException('Failing by intention');
     }
